requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
schedule==1.2.1
python-fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
pandas==2.2.1
PyYAML==6.0.1
python-dateutil==2.8.2
asyncio==3.4.3
python-dotenv==1.0.1 